feat: migrate create-react-app to vite
