+++ package.json
@@ -1,51 +1,35 @@
 {
-  "name": "react-multi-page-website",
-  "version": "0.1.0",
+  "name": "dtft",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "concurrently --kill-others --names \"NODE,REACT\" -c \"bgGreen.black,bgCyan.black\" \"nodemon ./server.js\" \"vite build --watch --mode development\"",
+    "dev:api": "nodemon ./server.js",
+    "dev:ui": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "bootstrap": "^5.2.2",
-    "browser-router": "^0.2.0",
+    "bootstrap": "^5.2.3",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
-    "mongodb": "^4.11.0",
-    "node": "^19.0.1",
+    "mongodb": "^4.13.0",
     "react": "^18.2.0",
-    "react-bootstrap": "^2.6.0",
+    "react-bootstrap": "^2.7.0",
     "react-dom": "^18.2.0",
-    "react-icons": "^4.6.0",
-    "react-router-dom": "^6.4.3",
-    "react-scripts": "5.0.1",
-    "react-scroll": "^1.8.8",
-    "serve": "^14.1.2",
-    "styled-components": "^5.3.6",
-    "web-vitals": "^2.1.4"
+    "react-icons": "^4.7.1",
+    "react-router-dom": "^6.6.1",
+    "react-scroll": "^1.8.9",
+    "styled-components": "^5.3.6"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "devDependencies": {
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
+    "@vitejs/plugin-react": "^3.0.1",
+    "concurrently": "^7.6.0",
+    "nodemon": "^2.0.20",
+    "vite": "^4.0.4"
   }
 }
