+++ server.js
@@ -1,17 +1,20 @@
-require('dotenv').config()
-const express = require("express");
+import dotenv from 'dotenv';
+import path from 'path';
+dotenv.config({ path: path.resolve(path.join(process.cwd(), '.env')) });
+import express  from "express";
 const app = express(); //create new express app
 const port = process.env.PORT; //using designated port number on front end to access back end
 // run server on specific port(5001), and the front end/index on port(3000)
 
-const { MongoClient } = require("mongodb"); //create a new mongoDB client
+import { MongoClient } from "mongodb"; //create a new mongoDB client
 const url = process.env.MONGO_URI;
 const client = new MongoClient(url);
 
-app.use(express.static("build")); //allows you to pass json data from front end to back end
+app.use(express.static("dist")); //allows you to pass json data from front end to back end
 app.use(express.urlencoded({ extended: true })); //allows you to access req.body
 
-const cors = require("cors");
+import cors from 'cors';
+
 // cors is a pkg that lets you make request across different urls/different machines -- makes port 3000 talk with server.js port 3001
 
 app.use(cors());
@@ -135,7 +138,7 @@ app.get("/api", function (req, res) {
 }
 
 app.get("*", (req, res) => {
-  res.sendFile(__dirname + "/build/index.html");
+    res.sendFile(path.resolve(path.join('dist', 'index.html')));
 });
 
 app.listen(port, () =>{
