+++ src/Pages/Home.jsx
@@ -1,7 +1,7 @@
 import "bootstrap/dist/css/bootstrap.min.css";
-import NavToggle from "../components/Navigation/NavToggle.js";
-import TruckSelector from "../components/TruckSelector.js";
-import Footer from "../components/Footer.js";
+import NavToggle from "../components/Navigation/NavToggle.jsx";
+import TruckSelector from "../components/TruckSelector.jsx";
+import Footer from "../components/Footer.jsx";
 
 export default function Home() {
   return (
