+++ src/components/IndividualModal.jsx
