+++ src/components/TruckDisplay.jsx
